Update agents guide for module split focus
