+++ agents.md
@@ -2,6 +2,7 @@
 
 This document collects ideas and instructions for implementing future improvements. Follow these when adding features or refactoring the code.
 After reading this guide, consult `characters.md` for the current UTF-8 plan and progress.
+Refer to the module outline in [modules.md](modules.md) when moving code out of `ee.c`.
 
 ## Repository Clean‑up
 
@@ -67,4 +68,6 @@ This file serves as a to‑do list and style reference. Update it whenever a new
 
 ## Current Work
 
-UTF-8 buffer refactoring has begun. See `characters.md` for the ongoing plan.
+The priority is splitting the remaining logic in `ee.c` into dedicated modules.
+Follow the layout in [modules.md](modules.md). UTF‑8 refactoring from
+`characters.md` continues once the split is mostly done.
